finnhub_tf is still a filter, with a caveat.
It still works as a filter; that is, if you pass in an already finnhub valid timeframe, it will spit that same value back out. So, use it as a filter now because incremental is called directly by tests.

But - must make sure you check for True evaluation of the result.